LPS-55800 Move TemplateSelector tag
