+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-taglib/src/META-INF/resources/taglib/ui/template_selector/page.jsp
@@ -14,7 +14,7 @@
  */
 --%>
 
-<%@ include file="/html/taglib/ddm/template_selector/init.jsp" %>
+<%@ include file="/taglib/ui/template_selector/init.jsp" %>
 
 <%
 long classNameId = GetterUtil.getLong((String)request.getAttribute("ddm:template-selector:classNameId"));
